Move glossary interface to a separate file
