+++ libs/sdmx-toolkit/src/models/structural-metadata/glossary.ts
@@ -1,0 +1,9 @@
+import { CodelistItemBase } from './hierarchy';
+
+export interface Glossary {
+  id: string;
+  agencyID?: string;
+  version?: string;
+  name?: string;
+  terms?: CodelistItemBase[];
+}
